+++ core/src/org/labkey/core/webdav/DavController.java
@@ -82,7 +82,9 @@ import org.labkey.api.view.ViewContext;
 import org.labkey.api.view.ViewServlet;
 import org.labkey.api.view.WebPartView;
 import org.labkey.api.view.template.PageConfig;
+import org.labkey.api.webdav.CaseSensitiveDavPath;
 import org.labkey.api.webdav.DavException;
+import org.labkey.api.webdav.DavPath;
 import org.labkey.api.webdav.DirectRequest;
 import org.labkey.api.webdav.WebdavResolver;
 import org.labkey.api.webdav.WebdavResolverImpl;
@@ -5603,41 +5605,8 @@ public class DavController extends SpringActionController
         return resolvePath(path, reload);
     }
 
-
-    public static class DavPath extends Path
-    {
-        public DavPath(Path path)
-        {
-            super(path);
-        }
-
-        @Override
-        protected int compareName(String a, String b)
-        {
-            if (FileUtil.isCaseInsensitiveFileSystem())
-                return a.compareToIgnoreCase(b);
-
-            return a.compareTo(b);
-        }
-    }
-
-    public static class S3Path extends Path
-    {
-        public S3Path(Path path)
-        {
-            super(path);
-        }
-
-        @Override
-        protected int compareName(String a, String b)
-        {
-            return a.compareTo(b); // S3 files are always case-sensitive
-        }
-    }
-
     // per request cache
     Map<DavPath, WebdavResolver.LookupResult> resourceCache = new HashMap<>();
-    Map<S3Path, WebdavResolver.LookupResult> s3ResourceCache = new HashMap<>();
     WebdavResolver.LookupResult nullDavFileInfo = new WebdavResolver.LookupResult(null,null);
 
     @Nullable WebdavResource resolvePath(String path)
@@ -5663,24 +5632,23 @@ public class DavController extends SpringActionController
 
         if (!reload)
         {
-            result = s3ResourceCache.get(new S3Path(path));
+            result = resourceCache.get(new CaseSensitiveDavPath(path)); // first try to find by strict case match
             if (result == null)
-                result = resourceCache.get(new DavPath(path));
+            {
+                result = resourceCache.get(new DavPath(path)); // then try os-based case match
+                if (result != null && result.resource != null && result.resource.toDavPath(path) instanceof CaseSensitiveDavPath)
+                    result = null; // reject cache if cache key class doesn't match
+            }
+
         }
 
         if (result == null)
         {
             result = getResolver().lookupEx(path);
-            DavPath davPath = new DavPath(path);
             if (result == null || result.resource == null)
-                resourceCache.put(davPath, nullDavFileInfo);
+                resourceCache.put(new DavPath(path), nullDavFileInfo);
             else
-            {
-                if (result.resource.isCloudResource()) // S3 files are always case-sensitive
-                    s3ResourceCache.put(new S3Path(path), result);
-                else
-                    resourceCache.put(davPath, result); // Otherwise make a best guess based on server OS
-            }
+                resourceCache.put(result.resource.toDavPath(path), result);
         }
 
         if (null == result || nullDavFileInfo == result || null == result.resource)
@@ -6675,10 +6643,7 @@ public class DavController extends SpringActionController
             isFile = r.getFile().isFile();
             if (isFile)
             {
-                if (r.isCloudResource())
-                    s3ResourceCache.remove(new S3Path(r.getPath()));
-                else
-                    resourceCache.remove(new DavPath(r.getPath()));
+                resourceCache.remove(r.toDavPath(r.getPath()));
                 r = resolvePath(r.getPath());
                 isFile = r.isFile();
             }
